refactor: Move authorize logics to Interceptor
