+++ src/main/java/nextstep/app/interceptor/AuthenticationInterceptor.java
@@ -1,0 +1,49 @@
+package nextstep.app.interceptor;
+
+import nextstep.app.exception.AuthenticationException;
+import nextstep.app.domain.MemberService;
+import nextstep.app.support.SecurityContextHolder;
+import org.springframework.http.HttpHeaders;
+import org.springframework.stereotype.Component;
+import org.springframework.util.Base64Utils;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@Component
+public class AuthenticationInterceptor implements HandlerInterceptor {
+
+    private final MemberService memberService;
+
+    public AuthenticationInterceptor(MemberService memberService) {
+        this.memberService = memberService;
+    }
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
+        this.authorize(request);
+
+        return true;
+    }
+
+    private void authorize(HttpServletRequest request) {
+        String authorization = request.getHeader(HttpHeaders.AUTHORIZATION);
+
+        String[] split = authorization.split(" ");
+        String type = split[0];
+        String credential = split[1];
+
+        if (!"Basic".equalsIgnoreCase(type)) {
+            throw new AuthenticationException();
+        }
+
+        String decodedCredential = new String(Base64Utils.decodeFromString(credential));
+        String[] emailAndPassword = decodedCredential.split(":");
+
+        String email = emailAndPassword[0];
+        String password = emailAndPassword[1];
+
+        memberService.isValidMemberInfo(email, password);
+    }
+}
